fix pcg ci
