--- test/registered/piecewise_cuda_graph/test_piecewise_cuda_graph_2_gpu.py
@@ -1,60 +1,0 @@
-import unittest
-
-from sglang.srt.utils import kill_process_tree
-from sglang.test.ci.ci_register import register_cuda_ci
-from sglang.test.run_eval import run_eval
-from sglang.test.test_utils import (
-    DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-    DEFAULT_URL_FOR_TEST,
-    CustomTestCase,
-    SimpleNamespace,
-    popen_launch_server,
-)
-
-# CI Registration - 2-GPU tests (80GB GPUs required)
-register_cuda_ci(est_time=160, suite="stage-b-test-large-2-gpu")
-
-
-class TestPiecewiseCudaGraphTP(CustomTestCase):
-    """Test piecewise CUDA graph with normal TP"""
-
-    @classmethod
-    def setUpClass(cls):
-        cls.model = "Qwen/Qwen3-Coder-30B-A3B-Instruct"
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=[
-                "--piecewise-cuda-graph-compiler",
-                "eager",
-                "--tp",
-                "2",
-            ],
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_gsm8k_accuracy(self):
-        """Test GSM8K accuracy with 8-shot setting"""
-        num_examples = 2000
-
-        args = SimpleNamespace(
-            base_url=self.base_url,
-            model=self.model,
-            eval_name="mgsm_en",
-            num_examples=num_examples,
-            num_threads=min(num_examples, 1024),
-        )
-
-        metrics = run_eval(args)
-        print(f"GSM8K Accuracy: {metrics['score']:.3f}")
-
-        self.assertGreaterEqual(metrics["score"], 0.90)
-
-
-if __name__ == "__main__":
-    unittest.main()
